+++ assets/test.ps1
@@ -7,5 +7,8 @@ function ThrowOnNativeFailure {
 
 $env:PYO3_PYTHON = "python"
 
+cargo check --package=rsl-cli
+ThrowOnNativeFailure
+
 cargo run
 ThrowOnNativeFailure
