+++ crate/rsl-cli/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "rsl-cli"
+version = "6.0.3"
+authors = [
+    "Matthew Kirby <xoparr@gmail.com>",
+    "Fenhl <fenhl@fenhl.net>"
+]
+edition = "2018"
+
+[dependencies]
+serde_json = "1"
+
+[dependencies.derive_more]
+version = "0.99"
+default-features = false
+features = ["from"]
+
+[dependencies.rsl]
+path = "../rsl"
+
+[dependencies.structopt]
+version = "0.3"
+features = ["paw"]
+
+[dependencies.wheel]
+git = "https://github.com/fenhl/wheel"
+branch = "main"
