+++ crate/rsl-gui/src/main.rs
@@ -1,19 +1,9 @@
-#![deny(rust_2018_idioms, unused, unused_import_braces, unused_qualifications, warnings)]
+#![deny(rust_2018_idioms, unused, unused_import_braces, unused_qualifications, unused_crate_dependencies, warnings)]
 
 #![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]
 
 use {
-    std::{
-        fmt,
-        io::{
-            self,
-            Cursor,
-        },
-        path::PathBuf,
-        sync::Arc,
-    },
-    derive_more::From,
-    directories::ProjectDirs,
+    std::fmt,
     enum_iterator::IntoEnumIterator,
     iced::{
         Application,
@@ -35,34 +25,17 @@ use {
         },
         window,
     },
-    rand::{
-        distributions::WeightedError,
-        prelude::*,
-    },
-    serde::Serialize,
     smart_default::SmartDefault,
-    tokio::{
-        fs::File,
-        io::{
-            AsyncReadExt,
-            AsyncWriteExt,
-        },
-    },
-    zip::{
-        ZipArchive,
-        result::ZipError,
-    },
     rsl::{
-        NUM_RANDO_RANDO_TRIES,
-        NUM_TRIES_PER_SETTINGS,
-        Weights,
+        GenError,
+        GenOptions,
+        Preset,
+        PresetOptions,
     },
     crate::file::FilePicker,
 };
 
 mod file;
-
-ootr::uses!();
 
 #[derive(Debug, Clone)]
 enum Message {
@@ -126,16 +99,6 @@ impl fmt::Display for Tab {
     }
 }
 
-#[derive(SmartDefault, Clone, Copy)]
-struct Options {
-    #[default = true]
-    standard_tricks: bool,
-    #[default = true]
-    rsl_tricks: bool,
-    #[default = true]
-    random_starting_items: bool,
-}
-
 #[derive(SmartDefault)]
 enum GenState {
     #[default]
@@ -172,7 +135,7 @@ struct App {
     #[default(FilePicker::new(format!("Output Directory"), Message::ChangeOutputDir, Message::BrowseOutputDir))]
     output_dir: FilePicker<file::Folder, Message>,
     tab: Tab,
-    options: Options,
+    options: PresetOptions,
     gen: GenState,
 }
 
@@ -198,10 +161,14 @@ impl Application for App {
                 self.gen = GenState::Generating;
                 let base_rom = self.base_rom.data.as_ref().expect("generate button should be disabled if no base rom is given").clone();
                 let output_dir = self.output_dir.data.as_ref().expect("generate button should be disabled if no output dir is given").clone();
-                let tab = self.tab;
-                let options = self.options;
+                let options = match self.tab {
+                    Tab::League => GenOptions::League,
+                    Tab::Solo => GenOptions::Preset { preset: Preset::Solo, options: PresetOptions { world_count: 1, ..self.options } },
+                    Tab::CoOp => GenOptions::Preset { preset: Preset::CoOp, options: PresetOptions { world_count: 1, ..self.options } },
+                    Tab::Multiworld => GenOptions::Preset { preset: Preset::Multiworld, options: self.options },
+                };
                 return async move {
-                    match generate(base_rom, output_dir, tab, options).await {
+                    match rsl::generate(base_rom, output_dir, options).await {
                         Ok(()) => Message::SeedDone, //TODO button to open output dir
                         Err(e) => Message::GenError(e),
                     }
@@ -239,152 +206,19 @@ impl Application for App {
                     //TODO conditionals toggle?
                     .push(Checkbox::new(self.options.random_starting_items, "Randomize Starting Items", Message::ToggleRandomStartingItems))
                     //TODO world count (Multiworld only)
+                    .spacing(16)
                     .into(),
             })
             .push(Space::with_height(Length::Fill))
             .push(self.gen.view(disabled_reason))
+            .spacing(16)
+            .padding(16)
             .into()
     }
 }
 
 async fn check_for_updates() -> Message {
     Message::UpdateCheckComplete(false) //TODO
-}
-
-#[derive(Serialize)]
-enum CompressRom {
-    Patch,
-}
-
-#[derive(Serialize)]
-struct RandoSettings {
-    rom: PathBuf,
-    output_dir: PathBuf,
-    enable_distribution_file: bool,
-    distribution_file: PathBuf,
-    create_spoiler: bool,
-    create_cosmetics_log: bool,
-    compress_rom: CompressRom,
-}
-
-impl RandoSettings {
-    fn new(rom_path: impl Into<PathBuf>, distribution_path: impl Into<PathBuf>, output_dir: impl Into<PathBuf>) -> RandoSettings {
-        RandoSettings {
-            rom: rom_path.into(),
-            output_dir: output_dir.into(),
-            enable_distribution_file: true,
-            distribution_file: distribution_path.into(),
-            create_spoiler: true,
-            create_cosmetics_log: false,
-            compress_rom: CompressRom::Patch,
-        }
-    }
-}
-
-#[derive(Debug, From, Clone)]
-enum GenError {
-    Io(Arc<io::Error>),
-    Json(Arc<serde_json::Error>),
-    MissingHomeDir,
-    Reqwest(Arc<reqwest::Error>),
-    TriesExceeded,
-    #[from]
-    Weights(WeightedError),
-    Zip(Arc<ZipError>),
-}
-
-macro_rules! from_arc {
-    ($($from:ty => $to:ty, $variant:ident,)*) => {
-        $(
-            impl From<$from> for $to {
-                fn from(e: $from) -> $to {
-                    <$to>::$variant(Arc::new(e))
-                }
-            }
-        )*
-    };
-}
-
-from_arc! {
-    io::Error => GenError, Io,
-    serde_json::Error => GenError, Json,
-    reqwest::Error => GenError, Reqwest,
-    ZipError => GenError, Zip,
-}
-
-impl fmt::Display for GenError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            GenError::Io(e) => write!(f, "I/O error: {}", e),
-            GenError::Json(e) => write!(f, "JSON error: {}", e),
-            GenError::MissingHomeDir => write!(f, "failed to locate home directory"),
-            GenError::Reqwest(e) => if let Some(url) = e.url() {
-                write!(f, "HTTP error at {}: {}", url, e)
-            } else {
-                write!(f, "HTTP error: {}", e)
-            },
-            GenError::TriesExceeded => write!(f, "{} settings each tried {} times, all failed", NUM_RANDO_RANDO_TRIES, NUM_TRIES_PER_SETTINGS),
-            GenError::Weights(e) => e.fmt(f),
-            GenError::Zip(e) => e.fmt(f),
-        }
-    }
-}
-
-async fn generate(base_rom: impl Into<PathBuf>, output_dir: impl Into<PathBuf>, tab: Tab, options: Options) -> Result<(), GenError> {
-    let project_dirs = ProjectDirs::from("net", "Fenhl", "RSL").ok_or(GenError::MissingHomeDir)?;
-    let cache_dir = project_dirs.cache_dir();
-    let distribution_path = cache_dir.join("plando.json");
-    // ensure the correct randomizer version is installed
-    let rando_path = cache_dir.join(if tab == Tab::League { "ootr-league" } else { "ootr-latest" });
-    let repo_ref = if tab == Tab::League { LEAGUE_COMMIT_HASH } else { "Dev-R" };
-    if rando_path.join("version.py").exists() {
-        let mut version_string = String::default();
-        File::open(rando_path.join("version.py")).await?.read_to_string(&mut version_string).await?;
-        if tab == Tab::League {
-            if version_string.trim() != format!("__version__ = '{}'", LEAGUE_VERSION) {
-                tokio::fs::remove_dir_all(&rando_path).await?;
-            }
-        } else {
-            //TODO check and warn for outdated versions
-        }
-    }
-    if !rando_path.exists() {
-        let rando_download = reqwest::get(&format!("https://github.com/Roman971/{}/archive/{}.zip", REPO_NAME, repo_ref)).await?
-            .bytes().await?;
-        ZipArchive::new(Cursor::new(rando_download))?.extract(&cache_dir)?; //TODO async
-        tokio::fs::rename(cache_dir.join(format!("{}-{}", REPO_NAME, repo_ref)), &rando_path).await?;
-    }
-    // write base rando settings to a file to be used as parameter later
-    let buf = serde_json::to_vec_pretty(&RandoSettings::new(base_rom, &distribution_path, output_dir))?; //TODO async-json
-    let settings_path = cache_dir.join("settings.json");
-    File::create(&settings_path).await?.write_all(&buf).await?;
-    // generate seed
-    let mut weights = serde_json::from_str::<Weights>(include_str!("../../../assets/weights/rsl.json"))?; //TODO allow custom weights
-    match tab {
-        Tab::League | Tab::Solo => {} // no modifications
-        Tab::CoOp => weights += serde_json::from_str(include_str!("../../../assets/weights/override-coop.json"))?,
-        Tab::Multiworld => weights += serde_json::from_str(include_str!("../../../assets/weights/override-multiworld.json"))?, //TODO apply world count
-    }
-    if tab != Tab::League { //TODO `&& tab != Tab::Custom`
-        match (options.standard_tricks, options.rsl_tricks) {
-            (true, true) => {}
-            (true, false) => weights.allowed_tricks = Some(serde_json::from_str(include_str!("../../../assets/weights/tricks-standard.json"))?),
-            (false, true) => weights.allowed_tricks = Some(serde_json::from_str(include_str!("../../../assets/weights/tricks-rsl.json"))?),
-            (false, false) => weights.allowed_tricks = Some(Vec::default()),
-        }
-        if !options.random_starting_items { weights.random_starting_items = false }
-    }
-    #[cfg(unix)] let python = "python3";
-    #[cfg(all(windows, debug_assertions))] let python = "python";
-    #[cfg(all(windows, not(debug_assertions)))] let python = "pythonw";
-    for _ in 0..NUM_RANDO_RANDO_TRIES {
-        let buf = serde_json::to_vec_pretty(&weights.gen(&mut thread_rng())?)?; //TODO async-json
-        File::create(&distribution_path).await?.write_all(&buf).await?;
-        for _ in 0..NUM_TRIES_PER_SETTINGS {
-            if tokio::process::Command::new(python).arg("OoTRandomizer.py").arg("--settings").arg(&settings_path).current_dir(&rando_path).status().await?.success() { return Ok(()) }
-        }
-    }
-    Err(GenError::TriesExceeded)
 }
 
 fn main() -> iced::Result {
