+++ crate/rsl/Cargo.toml
@@ -8,14 +8,32 @@ authors = [
 edition = "2018"
 
 [dependencies]
+directories = "3"
 enum-iterator = "0.6"
 rand = "0.8"
 rand_distr = "0.4"
 serde_json = "1"
+smart-default = "0.6"
+structopt = "0.3"
+zip = "0.5"
+
+[dependencies.derive_more]
+version = "0.99"
+default-features = false
+features = ["from"]
 
 [dependencies.ootr]
 path = "../ootr"
 
+[dependencies.reqwest]
+version = "0.10" #TODO update to 0.11 once iced supports tokio 1.0
+default-features = false
+features = ["rustls-tls"]
+
 [dependencies.serde]
 version = "1"
 features = ["derive"]
+
+[dependencies.tokio]
+version = "0.2" #TODO upgrade to 1 once supported by iced
+features = ["fs", "process"]
